backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
langchain
langchain-community
langchain-openai
langchain-core
langchain-text-splitters
chromadb
pypdf
docx2txt
openai
